movies_fastapi/requirements.txt
aioredis==2.0.0
elasticsearch[async]==7.9.1
fastapi==0.61.1
orjson==3.8.2
pydantic==1.9.0
uvicorn==0.12.2
# uvloop==0.14.0 На винде не ставится.
python-dotenv==0.21.0
fastapi-cache2==0.1.9
fastapi-cache2[redis]==0.1.9
gunicorn==20.1.0